requirements.txt
requests==2.25.1
biopython==1.78